refactor(xarray,utils): require integer dtype for k/j/i voxel coordinates
k/j/i are discrete voxel indices and were force-cast to float64 in
attach_voxel_to_world_index, purely as an implementation simplification —
nothing downstream actually needs fractional voxel positions (sel() always
rounds to nearest integer position before indexing). Enforce integer dtype
instead, and normalize storage to int64.

Fixes every construction site that built k/j/i with an explicit float dtype
(create_fusi_dataarray, the AUTC Zarr writer, the BrainGlobe atlas builder,
multipose consolidation, reindex_voxels, atlas resample/resample_like), and
updates test fixtures that used float literals for what were always integer
voxel coordinates.

+++ src/confusius/atlas/_accessor.py
@@ -478,10 +478,7 @@ class AtlasAccessor:
             `"i"`.
         """
         dims = [str(dim) for dim in dims]
-        coords = {
-            dim: np.arange(size, dtype=float)
-            for dim, size in zip(dims, shape, strict=True)
-        }
+        coords = {dim: np.arange(size) for dim, size in zip(dims, shape, strict=True)}
         reference = xr.DataArray(
             np.zeros(tuple(shape), dtype=np.float32),
             dims=dims,
